+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Added
+
+- [Godog](https://github.com/DATA-DOG/godog) runner configuration
+
+
 ## [0.10.0] - 2018-01-18
 
 ### Removed